Add files via upload
This repository presents a comprehensive analysis of the unemployment trends in India during the COVID-19 pandemic, using R for all aspects of data processing, analysis, and visualization. The project delves into the intricacies of unemployment rates across different Indian states, with a focus on urban and rural disparities and the broader economic implications of the pandemic. Contained within are the original datasets, R scripts for cleaning and analysis, and a series of visualizations that encapsulate the project's findings. It serves as a resource for policymakers, academics, and social scientists interested in the socioeconomic effects of COVID-19 on the Indian labor market.